ml/pyproject.toml
[project]
name = "falldet"
version = "0.1.0"
description = "ML training infrastructure for IMU-based fall detection"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.2",
    "numpy>=1.26",
    "scipy>=1.12",
    "pandas>=2.2",
    "scikit-learn>=1.4",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "h5py>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "jupyter>=1.0",
    "ipykernel>=6.29",
    "ruff>=0.3",
]
tracking = [
    "mlflow>=2.11",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["falldet*"]

[tool.ruff]
line-length = 100
